pyproject.toml
[project]
name = "ckanext-versioned-tiledmap"
version = "2.3.2"
description = "A CKAN extension with a map view for versioned-datastore backed resources."
readme = "README.md"
requires-python = ">=3.6"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Natural History Museum", email = "data@nhm.ac.uk" }
]
keywords = ["CKAN", "data", "versioned_tiledmap"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]
dependencies = [
    "ckanext-versioned-datastore>=3.7.0",
    "cachetools>=4.2.4",
    "ckantools>=0.4.1"
]

[project.optional-dependencies]
test = [
    "mock",
    "pytest>=4.6.5",
    "pytest-cov>=2.7.1",
    "coveralls"
]

[project.urls]
repository = "https://github.com/NaturalHistoryMuseum/ckanext-versioned-tiledmap"
changelog = "https://github.com/NaturalHistoryMuseum/ckanext-versioned-tiledmap/blob/main/CHANGELOG.md"

[project.entry-points."ckan.plugins"]
versioned_tiledmap = "ckanext.tiledmap.plugin:VersionedTiledMapPlugin"


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[tool]
[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
exclude = ["tests", "docs"]

[tool.setuptools.package-data]
"ckanext.tiledmap.theme" = ["*", "**/*"]

[tool.commitizen]
name = "cz_nhm"
version = "2.3.2"
tag_format = "v$version"
update_changelog_on_bump = true
changelog_incremental = true
version_files = [
    "pyproject.toml:version",
    "CITATION.cff:^version"
]

[tool.ruff.format]
quote-style = "single"

[tool.pylint]
max-line-length = 88
disable = ["C0114", "R0903"]

[tool.docformatter]
wrap-summaries = 88
wrap-descriptions = 88
pre-summary-newline = true
make-summary-multi-line = true
